+++ src/reavers.c
@@ -1252,12 +1252,12 @@ void apply_reaver_mods(P_char ch)
     if((w1 && required_weapon_skill(w1) == SKILL_1H_BLUDGEON) ||
        (w2 && required_weapon_skill(w2) == SKILL_1H_BLUDGEON))
     {
-      ch->specials.base_combat_round += (int) get_property("reavers.ice.penalty.pulse.1hBludgeon", 2);
+      ch->specials.base_combat_round += get_property("reavers.ice.penalty.pulse.1hBludgeon", 2.);
     }
     else if((w1 && required_weapon_skill(w1) == SKILL_2H_BLUDGEON) ||
             (w2 && required_weapon_skill(w2) == SKILL_2H_BLUDGEON))
     {
-      ch->specials.base_combat_round += (int) get_property("reavers.ice.penalty.pulse.2hBludgeon", 3);
+      ch->specials.base_combat_round += get_property("reavers.ice.penalty.pulse.2hBludgeon", 3.);
     }
 
   }
@@ -1268,7 +1268,7 @@ void apply_reaver_mods(P_char ch)
        required_weapon_skill(w1) == SKILL_2H_FLAYING ||
        required_weapon_skill(w2) == SKILL_2H_FLAYING )
     {
-      ch->specials.base_combat_round += (int) get_property("reavers.flame.penalty.pulse.2hSlashing", 3);
+      ch->specials.base_combat_round += get_property("reavers.flame.penalty.pulse.2hSlashing", 3.);
     }
   }
 /* Shock reavers currently have no two handed weapon skills. Apr09 -Lucrot
@@ -1277,7 +1277,7 @@ void apply_reaver_mods(P_char ch)
     if( required_weapon_skill(w1) == SKILL_2H_SLASHING ||
         required_weapon_skill(w2) == SKILL_2H_SLASHING )
     {
-      ch->specials.base_combat_round += (int) get_property("reavers.shock.penalty.pulse.2hSlashing", 3);
+      ch->specials.base_combat_round += get_property("reavers.shock.penalty.pulse.2hSlashing", 3.);
     }
   } */
 }
